use shared_ptr instead of unique_ptr to overcome msvc compiler error. Fix/hide a few warnings

+++ pdal/libpdalpython.pyx
@@ -97,7 +97,7 @@ cdef class PyPipeline:
 
     def __cinit__(self, unicode json, list arrays=None):
         cdef char* x = NULL
-        cdef int n_arrays;
+        cdef Py_ssize_t n_arrays;
         if arrays:
             n_arrays = len(arrays)
